requirements.txt
python-telegram-bot
requests
python-dotenv
sqlalchemy
psycopg2-binary
selenium
webdriver-manager